Reorganizes all helper functions (READ MORE)
- Adds new file, helpers.js
- Adds all helper functions from nudge and message controllers to helpers
- Adds all modules and constants needed to run those functions to helpers
- Removes modules and functions from message and nudge controllers.
- Imports helpers.js into message and nudge controllers.
- adjusts routes using those functions to correct invocation syntax.
